#205 Add functional test covering package installation & removal
