+++ phpstan.neon.dist
@@ -8,6 +8,7 @@ parameters:
       - Lemberg\Draft\Environment\Config\Manager\AbstractConfigManager
       - Lemberg\Draft\Environment\Config\Install\Step\AbstractInstallStep
       - Lemberg\Draft\Environment\Config\Update\Step\AbstractUpdateStep
+      - Lemberg\Tests\Functional\Draft\Environment\AbstractFunctionalTest
       - Lemberg\Tests\Functional\Draft\Environment\Config\Manager\AbstractConfigManagerTest
       - Symfony\Component\Filesystem\Filesystem
   ignoreErrors:
@@ -15,4 +16,3 @@ parameters:
       message: '#Method [A-Za-z0-9\\]+::[A-Za-z0-9\(\)]+ is not final, but since the containing class is abstract, it should be.#'
       paths:
         - src/Config/Install/Step/AbstractInstallStep.php
-        - tests/Functional/Config/Manager/AbstractConfigManagerTest.php
